{"types":"C","runs":17}

+++ problem16.java
@@ -6,7 +6,7 @@ public class problem16 extends ConsoleProgram {
 		int firstNum = readInt("First number is: ");
 		int secondNum = readInt("Second number is: ");
 		int thirdNum = readInt("Third number is : ");
-		double average = 1 / (1.0 / firstNum + 1.0 / secondNum + 1.0 / thirdNum);
+		double average = 1 / (1.0 / firstNum + 1 / (double) secondNum + 1.0 / thirdNum);
 		println("Average is: " + average);		
 	}
 }
